requirements.txt
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.2.0
exceptiongroup==1.1.2
h11==0.14.0
idna==3.4
lxml==4.9.3
outcome==1.2.0
packaging==23.1
psycopg2-binary==2.9.7
PySocks==1.7.1
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
selenium==4.14.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
tqdm==4.65.0
trio==0.22.2
trio-websocket==0.10.3
urllib3==2.0.3
webdriver-manager==3.8.6
wsproto==1.2.0
